+++ pandapower/file_io.py
@@ -87,7 +87,7 @@ def to_excel(net, filename, include_empty_tables=False, include_results=True):
     writer.save()
 
 
-def to_json(net, filename=None, encryption_key=None):
+def to_json(net, filename=None, encryption_key=None, store_index_names=False):
     """
         Saves a pandapower Network in JSON format. The index columns of all pandas DataFrames will
         be saved in ascending order. net elements which name begins with "_" (internal elements)
@@ -97,11 +97,15 @@ def to_json(net, filename=None, encryption_key=None):
             **net** (dict) - The pandapower format network
 
             **filename** (string or file, None) - The absolute or relative path to the output file
-                                                  or a file-like object,
-                                                  if 'None' the function returns a json string
+            or a file-like object, if 'None' the function returns a json string
 
             **encrytion_key** (string, None) - If given, the pandapower network is stored as an
-                                               encrypted json string
+            encrypted json string
+
+            **store_index_names** (bool, False) - If True, an additional dict "index_names" is
+            stored into the json string which includes the index names of the dataframes within the
+            net.
+            Since pandapower does usually not use net[elm].index.name, the default is False.
 
 
         EXAMPLE:
@@ -109,9 +113,26 @@ def to_json(net, filename=None, encryption_key=None):
              >>> pp.to_json(net, "example.json")
 
     """
+    # --- store index names
+    if store_index_names:
+        # To ensure correct index names (see https://github.com/e2nIEE/pandapower/issues/1410),
+        # these are additionally stored to the json file as a dict.
+        if "index_names" in net.keys():
+            raise ValueError("To store DataFrame index names, 'index_names' "
+                             "is used and thus should not be a key of net.")
+        net["index_names"] = {
+            key: net[key].index.name for key in net.keys() if isinstance(
+                net[key], pd.DataFrame) and isinstance(net[key].index.name, str) and \
+                net[key].index.name != ""
+        }
+
     json_string = json.dumps(net, cls=io_utils.PPJSONEncoder, indent=2)
     if encryption_key is not None:
         json_string = io_utils.encrypt_string(json_string, encryption_key)
+
+    if store_index_names:
+        # remove the key "index_names" to not change net
+        del net["index_names"]
 
     if filename is None:
         return json_string
@@ -224,7 +245,8 @@ def _from_excel_old(xls):
 
 
 def from_json(filename, convert=True, encryption_key=None, elements_to_deserialize=None,
-              keep_serialized_elements=True, add_basic_std_types=False, replace_elements=None):
+              keep_serialized_elements=True, add_basic_std_types=False, replace_elements=None,
+              empty_dict_like_object=None, restore_index_names=True):
     """
     Load a pandapower network from a JSON file.
     The index of the returned network is not necessarily in the same order as the original network.
@@ -248,8 +270,16 @@ def from_json(filename, convert=True, encryption_key=None, elements_to_deseriali
         **add_basic_std_types** (bool, False) - Add missing standard-types from pandapower standard
         type library.
 
-        **replace_elements** (dict, None) - Keys are replaced by values found in json string. Both key and
-        value are supposed to be strings.
+        **replace_elements** (dict, None) - Keys are replaced by values found in json string.
+        Both key and value are supposed to be strings.
+
+        **empty_dict_like_object** (dict/pandapower.pandapowerNet/..., None) - If None,
+        the output of pandapower.create_empty_network() is used as an empty element to be filled by
+        the data of the json string. Give another dict-like object to start filling that alternative
+        object with the json data.
+
+        **restore_index_names** (bool, True) - If True, the index names of dataframes in the net
+        are restored using information of a dict "index_names" (if this dict is available).
 
     OUTPUT:
         **net** (dict) - The pandapower format network
@@ -267,14 +297,21 @@ def from_json(filename, convert=True, encryption_key=None, elements_to_deseriali
         with open(filename, "r") as fp:
             json_string = fp.read()
 
-    return from_json_string(json_string, convert=convert, encryption_key=encryption_key,
-                            elements_to_deserialize=elements_to_deserialize,
-                            keep_serialized_elements=keep_serialized_elements,
-                            add_basic_std_types=add_basic_std_types, replace_elements=replace_elements)
+    return from_json_string(
+        json_string,
+        convert=convert,
+        encryption_key=encryption_key,
+        elements_to_deserialize=elements_to_deserialize,
+        keep_serialized_elements=keep_serialized_elements,
+        add_basic_std_types=add_basic_std_types,
+        replace_elements=replace_elements,
+        empty_dict_like_object=empty_dict_like_object,
+        restore_index_names=restore_index_names)
 
 
 def from_json_string(json_string, convert=False, encryption_key=None, elements_to_deserialize=None,
-                     keep_serialized_elements=True, add_basic_std_types=False, replace_elements=None):
+                     keep_serialized_elements=True, add_basic_std_types=False,
+                     replace_elements=None, empty_dict_like_object=None, restore_index_names=True):
     """
     Load a pandapower network from a JSON string.
     The index of the returned network is not necessarily in the same order as the original network.
@@ -297,8 +334,16 @@ def from_json_string(json_string, convert=False, encryption_key=None, elements_t
         **add_basic_std_types** (bool, False) - Add missing standard-types from pandapower standard
         type library.
 
-        **replace_elements** (dict, None) - Keys are replaced by values found in json string. Both key and
-        value are supposed to be strings.
+        **replace_elements** (dict, None) - Keys are replaced by values found in json string.
+        Both key and value are supposed to be strings.
+
+        **empty_dict_like_object** (dict/pandapower.pandapowerNet/..., None) - If None,
+        the output of pandapower.create_empty_network() is used as an empty element to be filled by
+        the data of the json string. Give another dict-like object to start filling that alternative
+        object with the json data.
+
+        **restore_index_names** (bool, True) - If True, the index names of dataframes in the net
+        are restored using information of a dict "index_names" (if this dict is available).
 
     OUTPUT:
         **net** (dict) - The pandapower format network
@@ -316,9 +361,11 @@ def from_json_string(json_string, convert=False, encryption_key=None, elements_t
         json_string = io_utils.decrypt_string(json_string, encryption_key)
 
     if elements_to_deserialize is None:
-        net = json.loads(json_string, cls=io_utils.PPJSONDecoder)
+        net = json.loads(json_string, cls=io_utils.PPJSONDecoder,
+                         empty_dict_like_object=empty_dict_like_object)
     else:
-        net = json.loads(json_string, cls=io_utils.PPJSONDecoder, deserialize_pandas=False)
+        net = json.loads(json_string, cls=io_utils.PPJSONDecoder, deserialize_pandas=False,
+                         empty_dict_like_object=empty_dict_like_object)
         net_dummy = create_empty_network()
         if ('version' not in net.keys()) | (version.parse(net.version) < version.parse('2.1.0')):
             raise UserWarning('table selection is only possible for nets above version 2.0.1. '
@@ -359,6 +406,14 @@ def from_json_string(json_string, convert=False, encryption_key=None, elements_t
         net_dummy = create_empty_network()
         for key in net_dummy.std_types:
             net.std_types[key] = dict(net_dummy.std_types[key], **net.std_types[key])
+    if restore_index_names and "index_names" in net.keys():
+        if not isinstance(net.index_names, dict):
+            raise ValueError("To restore the index names of the dataframes, a dict including this "
+                             f"information is expected, not {type(net.index_names)}")
+        for key, index_name in net["index_names"].items():
+            if key in net.keys():
+                net[key].index.name = index_name
+        del net["index_names"]
 
     return net