+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ Change Log
 [2.10.1] - 2022-07-31
 -------------------------------
 - [FIXED] remove the parameter ignore_order in DeepDiff (__eq__), add __hash__ to JSONSerializableClass
+- [ADDED] store and restore functionality of dataframe index names with to_json() and from_json()
+- [ADDED] generalization from_json() with parameter empty_dict_like_object
 
 [2.10.0] - 2022-07-29
 -------------------------------